[#22]; refactor: 패키지 위치 변경
